# 변경사항 1. BYE 1차 프로토타입 개발 완료 2. APPLICATION-SPECIFIC 개발 해야함 3. Header 에서 length 값을 바이트 수로 변환해야함

+++ framework/src/main/java/network/rtcp/type/base/RtcpReportBlock.java
@@ -80,7 +80,7 @@ public class RtcpReportBlock {
     // CONSTRUCTOR
     public RtcpReportBlock(long ssrc, short f, int c, long ehsn, long j, long lsr, long dlsr) {
         this.ssrc = (int) ssrc;
-        this.f = f;
+        this.f = (byte) f;
         this.c = c;
         this.ehsn = (int) ehsn;
         this.j = (int) j;
